+++ templates/privacy.html
@@ -1,0 +1,18 @@
+{% extends "layout.html" %}
+
+{% block title %}
+    Privacy
+{% endblock %}
+
+{% block main %}
+<div class="header">
+    <h1>Privacy</h1>
+</div>
+<div class="container">
+    <div class="section">
+        <h5>How does Cicero take care of my privacy</h5>
+        Cicero really cares about your privacy.
+    </div>  
+</div>
+
+{% endblock %}
