+++ templates/terms.html
@@ -1,0 +1,18 @@
+{% extends "layout.html" %}
+
+{% block title %}
+    Terms of service
+{% endblock %}
+
+{% block main %}
+<div class="header">
+    <h1>Terms of service</h1>
+</div>
+<div class="container">
+    <div class="section">
+        <h5>What are the terms of service?</h5>
+        Here are the terms of service.
+    </div>   
+</div>
+
+{% endblock %}
